+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 ### Added
 - `value::Opaque` can be used to wrap a Ruby type to make it `Send` + `Sync`.
 - `value::OpaqueId` is a `Send` + `Sync` version of `value::Id`.
+- `value::LazyId` is an `Id` with a `const` constructor so can be assigned to a
+  `static` and derefs to `OpaqueId`.
 - The `#[magnus(opaque_attr_reader)]` attribute can be set on `Opaque` wrapped
   fields of a struct when deriving `TypedData` to generate a method to return
   the inner value.